requirements.txt
Cython==0.29.4
Django==2.1.5
djangorestframework==3.9.1
numpy==1.16.1
opencv-python==4.0.0.21
tensorflow==1.12.0
Werkzeug==0.14.1